try_find_file is executed in the async thread
