+++ src/BabylonCpp/src/asio/internal/file_loader_sync.cpp
@@ -1,6 +1,6 @@
 #include <babylon/asio/internal/file_loader_sync.h>
 #include <babylon/asio/internal/sync_callback_runner.h>
-
+#include <babylon/core/filesystem.h>
 #include <fstream>
 
 namespace BABYLON {
@@ -8,11 +8,31 @@ namespace asio {
 namespace sync_io_impl
 {
 
+std::string tryFindFile(const std::string filename)
+{
+  // Check if the file is locally available
+  // - Check in local folder
+  auto absolutePath = Filesystem::absolutePath(filename);
+  if (Filesystem::exists(absolutePath)) {
+    return absolutePath;
+  }
+  // - Check in assets folder
+  absolutePath = Filesystem::absolutePath(BABYLON::assets_folder() + filename);
+  if (Filesystem::exists(absolutePath)) {
+    return absolutePath;
+  }
+
+  // Failure !
+  return "";
+};
+
 ArrayBufferOrErrorMessage LoadFileSync_Binary(
-  const std::string& filename,
+  const std::string& filenameNotSearched,
   const OnProgressFunction& onProgressFunction
 )
 {
+  std::string filename = tryFindFile(filenameNotSearched);
+
   std::ifstream ifs(filename.c_str(), std::ios::binary | std::ios::ate);
   if (!ifs.good())
     return ErrorMessage("LoadFileSync_Binary: Could not open file " + std::string(filename));
